Update Int32 to Double in specs

+++ src/specs/NativeEventBroadcastModule.ts
@@ -1,10 +1,10 @@
 /* eslint-disable @typescript-eslint/no-wrapper-object-types */
 import type { TurboModule } from 'react-native';
 import { TurboModuleRegistry } from 'react-native';
-import { Int32 } from 'react-native/Libraries/Types/CodegenTypes';
+import { Double } from 'react-native/Libraries/Types/CodegenTypes';
 
 export interface Spec extends TurboModule {
-  broadcastEvent(tag: Int32, event: Object): void;
+  broadcastEvent(tag: Double, event: Object): void;
 }
 
 // Note: codegen does not like `TurboModuleRegistry?.getEnforcing`
